Fix watchdog on Python3 (fixes #528)
A misunderstanding of the ioctl() call interface. If mutable=False then
fcntl.ioctl() actually returns the arg buffer back. This accidentally
worked on Python2 because int and str comparison did not return an error.
Error reporting is actually done by raising IOError on Python2 and OSError
on Python3.

Fix is to let IOError,OSError bubble up from LinuxWatchdog._ioctl and wrap
it in WatchdogError with error context information in calling functions.

Also properly handle errors in set_timeout(), have them result in only a
warning if watchdog support is not required.