Make it possible use endpoints for leader election instead of configmap
to enable it you have to set `kubernetes.use_endpoints` parameter.

If Patroni is using endpoints, it will use `cluster-name` endpoint
instead of `cluster-name-leader` for leader elections. Optionally
Patroni can update subsets for the Endpoint, by putting value of
`kubernetes.pod_ip` there, so the corresponding Service can be used for
directing a traffic to the master.