Expose members timeline and history in DCS
It's easy to get the timeline of the master by calling pg_xlogfile_name
and parsing its result. On the replica it requires to open a separate
replication connection. In order to avoid doing that on every iteration
of HA loop we will cache the result and re-evaluate it if it doesn't
match with the timeline of the leader.