Abort start if attaching to running postgres and cluster not initiazlied
Patroni can attach itself to an already running PostgreSQL instance. If
that is the first instance "seen" in the given cluster, Patroni for that
instance will create the initialize key, grab the leader key and, if the
instance is running a replica, promote.

Because of this behaviour, when a cluster with a master and one or more
replicas gets Patroni for each node, it is imperative to start running
Patroni on the master node before getting to the replicas.

This commit changes such weird behavior and will abort Patroni start if
there is no initialize key in DCS and postgres is running as replica.

Closes https://github.com/zalando/patroni/issues/655