Prevent tests from failing on Python 2
Python 2 does not have a list.copy() function. Implement our own
copy for lists and dictionaries (another option would be importing
copy module, but since we have only 2 types of objects to copy and
need nothing else from that module it is easier to just implement
the copy in Patroni).