BUGFIX: postmaster start can fail if pid from postmaster.pid is alive
Upon start postmaster process performs various safety checks if there is
a postmaster.pid file in the data directory. Although Patroni already
detected that the running process corresponding to the postmaster.pid is
not a postmaster, the new postmaster might fail to start, because it
thinks that postmaster.pid is already locked.
Important!!! Unlink of postmaster.pid isn't an option in this case,
because it has a lot of nasty race conditions.
Luckily there is a workaround to this problem, we can pass the pid from
postmaster.pid in the `PG_GRANDPARENT_PID` environment variable and
postmaster will ignore it.