Take and apply some parameters from controldata when starting as replica
https://www.postgresql.org/docs/10/static/hot-standby.html#HOT-STANDBY-ADMIN
There is set of parameters which value on the replica must be not
smaller than on the primary, otherwise replica will refuse to start:
* max_connections
* max_prepared_transactions
* max_locks_per_transaction
* max_worker_processes

It might happen that values of these parameters in the global
configuration are not set high enough, what makes impossible to start a
replica without human intervention. Usually it happens when we bootstrap
a new cluster from the basebackup.

As a solution to this problem we will take values of above parameters
from the pg_controldata output and in case if the values in the global
configuration are not high enough, apply values taken from
pg_controldata and set `pending_restart` flag.