Respond 200 to master health-check only if update_lock was successful
If Patroni gets partitioned it starts receiving stale information from
DCS. We can't use this information to determine that we have the leader
key. Instead, we will record in Ha object the actual state of
acquire/update lock and report as a leader only if it was successful.

P.S. despite responding with 200 on `GET /master` postgres was still
running read-only.