Stop-gap message of change or no change
patroni/postgresql.py:reload_config does the authoritative check on whether configuration options were changed or not. We only come here if the local configruation file was changed, so we have the short circuit message there. Otherwise we can get this far but if the values are actually the same then there will be no reload.