Implement support of permanent replication slots
Permanent replication slots are preserved on failover/switchover, that
is Patroni on the new primary will create configured replication slots
right after promote.

Slots could be configured with the help of `patronictl edit-config`.
The initial configuration could be also done in the `bootstrap.dcs`

```yaml
slots:
  permanent_physical_1:
    type: physical
  permanent_logical_1:
    type: logical
    database: foo
    plugin: pgoutput
```

It is responsibility of the operator to make sure that there are no
clashes in names between replication slots automatically created by
Patroni for members and permanent replication slots.