Do not skip `create_replica_methods`
In the original implementation of standby cluster all allowed properties
were checked if they're strings or not. But for `create_replica_methods`
it's wrong since it's a list, so this options is skipped. Fortunately,
if this options is not set, it will be taken from postgresql section,
but this issue needs to be fixed anyway.