Make it possible to automatically reinit the former master
If the pg_rewind is disabled or can't be used, the former master could
fail to start as a new replica due to diverged timelines. In this case
the only way to fix it is wiping the data directory and reinitilizing.

So far Patroni was able to remove the data directory only after failed
attempt to run pg_rewind. This commit fixes it.
If the `postgresql.remove_data_directory_on_diverged_timelines` is set,
Patroni will wipe the data directory and reinitialize the former master
automatically.

Fixes: https://github.com/zalando/patroni/issues/941