Adjust read timeout for leader watch blocking query
According to the Consul documentation the actual response timeout is
increased by a mall random amount of additional wait time added to the
supplied maximum wait time to spread out the wake up time of any
concurrent requests. It adds up to wait / 16 additional time to the
maximum duration.
In our case we will add wait/15 or 1 second depending on what is bigger.

Fixes: https://github.com/zalando/patroni/issues/945