Create headless service associated with $SCOPE-config endpoint
if there is no service defined k8s assumes that endpoint is orphaned and
removes it.
Patroni tries to create the service only in case if use_endpoints is
enabled if following cases:
1. Up on start
2. When it tries to (re-)create the config endpoint

If for some reason creation of the service has failed, Patroni will
retry it on every cycle of HA loop. Usually it fails due to lack of
permissions and if you don't want to give such permissions to the
service account used by Patroni, you can create the service explicitly
in the deployment manifest.