Change base port for acceptance tests from 5440 to 5340.
This allows for enough room in order not to interfere with any possibly running
host instances.  On Debian/Ubuntu, host instances get ports assigned starting
from 5432, so in the (rare but possible) case that 9 instances are running on
the host, the acceptance tests would fail to startup its instance.

Using 5340 as base port will allow for more than 90 instances to be created
during the acceptance tests, which should be plenty considering that currenlty
around 20-25 are created.