Better support for static etcd cluster
if the `etcd.use_proxies` is set to true, Patroni will stick to the list
of hosts specified in the `etcd.hosts` and avoid doing topology
discovery. Such mode might be useful when you know that you connect to
the etcd cluster via the set of proxies or when etcd cluster has static
topology.