Fix callbacks behavior (mostly for standby cluster)
`on_role_change` was never called for standby leader, only
`on_start`/`on_restart` and with a wrong role argument.

First of all, this patch changes behavior of `on_start`/`on_restart`
callbacks, they will be called only when postgres is started or
restarted without role changes. In case if the member is promoted or
demoted only the `on_role_change` callback will be executed.

In addition to that the REST API will return standby_leader role for the
leader of standby cluster.

Closes https://github.com/zalando/patroni/issues/988