Remove psycopg2 from requirements
Recently released psycopg2 split to two different packages, psycopg2 and
psycopg2-binary which could be installed at the same time into the same
place on the filesystem. In order to decrease dependency hell problem we
let a user choose how to install psycopg2. There are a few options
available and it is reflected in the documentation.

* `pip install patroni` will fail if psycopg2 is not installed
* Patroni will check psycopg2 up on start and fail if it can't be found
or outdated.