Add read-only and read-write API routes
Patroni is great behind a load balancer like HAProxy. API routes
are handy to direct the trafic to a valid node depending on its
role. We could have one route for writes directed to the primary
node and one route for reads directed to the replicas. In a two
nodes setup, when we loose a node, there's one host left: the
primary. Then, the read traffic will be dropped, even if the
primary can handle it. This commit adds read-only and read-write
routes. Read-only route enables reads on the primary. Read-write
route is an alias for '/master', '/primary' or '/leader' route.