Compatibility with postgresql 12
* use `SHOW primary_conninfo` instead of parsing config file on pg12
* parse the value of primary_conninfo in order to compare parameters
  instead of doing comparison of raw string
* cache parse results until the next change of config file (track mtime)
* strip out standby and recovery parameters from postgresql.auto.conf
  before starting the postgres 12
* simplify ConfigHandler.reload_config: the new compare_values function
  handles all corner cases for us

Patroni config remains backward compatible. Despite `restore_command`
converted to a GUC starting from postgresql 12, in the Patroni
configuration you can still keep it in the `postgresql.recovery_conf`
section.
If you put it into `postgresql.parameters.restore_command`, that will
also work, but it is important not to use both ways in the same time.