Implement two-step logging
A few times we observed that Patroni HA loop was blocked for a few
minutes due to not being able to write logs to stderr. This is a very
rare condition which we hit so far only on k8s. This commit makes
Patroni resilient to such kind of problems.
All log messages first are written into the in-memory queue and later
they are asynchronously flushed into the stderr or file from a separate
thread.

The maximum queue size is configurable and the default value is 1000.
This should be enough to keep more than one hour of log messages with
default settings and when Patroni cluster operates normally (without big
issues).

In case if we hit the maximum size of the queue further logs will be
discarded until the queue size will be reduced. The number of discarded
messages will be reported into the log later.

In addition to that, the number of non-flushed and discarded messages
(if there are any), will be reported via Patroni REST API as:
`logger_queue_size: X, logger_records_lost: Y`