Do a smart comparison of actual and desired primary_conninfo
On every replica Patroni periodically opens the recovery.conf file and
checks that it contains the correct primary_conninfo.
So far the correctness check was pretty dumb, it was basically doing a
full match of strings. That could lead to the restart of the replica
when Patroni is "joining" already running postgres process.
Instead of using string comparison we parse the actual primary_conninfo
value from the file and check that all parameters are matching with
the desired value.

In addition to that we stop reading and parsing recovery.conf of every
iteration if the modification time didn't change.