Add Etcd v3 protocol support via api grpc-gateway
The only python-etcd3 client working directly via gRPC still supports
only a single endpoint, what is not very nice for high-availability.

Since Patroni is already using a heavily hacked version of python-etcd
with smart retries and auto-discovery out-of-the-box, I decided to
enhance the existing code with a limited support of v3 protocol via
grpc-gateway.

Currently it passes all integration tests (behave). Unit tests are
missing.  Watch support is also missing. I.e. if the leader lease/key
disappeared from the Etcd, nodes will not be notified immediately, but
will learn about this fact only when running the next HA loop.
Effectively that means that the failover will be delayed up to
`loop_wait` seconds.
Implementing the watch support should not be hard, but due to a
streaming nature it will require a separate connection to the Etcd.
P.S. watch requests via grpc-gateway are broken on 3.0.X, but this is
probably not a big deal, since the current version is 3.4.0