Explicitly start logger Thread
The PatroniLogger object is instantiated in the Patroni constructor and
down the road there might be a fatal error causing Patroni process to
exit, but live thread prevents the normal shutdown.
In order to mitigate the issue and don't loose ability to use the
logging infrastructure we will switch to QueueLogger only when the
thread was explicitly started from the Patroni.run() method.
Continuation of https://github.com/zalando/patroni/pull/1178