Switch to a streaming watcher
Watch requests to K8s API either streaming the data or close connection
by timeout. In any case it requires a second connection open, but
opening a new connection every 10 seconds is more expensive for both,
Patroni and K8s API.

Switching to the streaming model also brings other benefits: we can
watch not only on leader object, but also on config, and wake up Patroni
main thread if the config was changed.