Perform pg_ctl reload regardless of config changes
It is possible that some config files are not controlled by Patroni and
when somebody is doing reload via REST API or by sending SIGHUP to
Patroni process the usual expectation is that postgres will also be
reloaded, but it didn't happen when there were no changes in postgresql
section of Patroni config.

For example one might replace ssl_cert_file and ssl_key_file and
starting from PostgreSQL 10 it just requires a reload, but Patroni
wasn't doing it.

In addition to that fix the issue with handling of `wal_buffers`.
The default value depends on `shared_buffers` and `wal_segment_size` and
therefore Patroni was exposing pending_restart when the new value in the
config was explicitly set to -1 (default).