Avoid calling expensive os.listdir()
When the cluster is under stress os.listdir() could take a few seconds
to execute what is badly affecting HA loop of Patroni and could even
cause the leader key to disappear from DCS due to the lack of updates.

There is a better and less expensive way to check that the PGDATA is not
empty. Instead of doing the listdir we simply check the presence of the
`global/pg_control` file in it.