Use LIST + WATCH when working with K8s API
There is an opinion that LIST requests with labelSelector are expensive
and Patroni was doing two such requests per HA loop (LIST pods and LIST
endpoints/configmaps).
In order to efficiently detect object changes we will switch to the
LIST+WATCH approach. The initial LIST request populates the ObjectCache
and later events from the WATCH request will update it. Also the
ObjectCache will be updated after performing the UPDATE or DELETE
operations on the K8s object. In order to avoid race conditions, all
operations on ObjectCache are performed after comparing the
resource_version of the old and the new objects and rejected if the new
resource_version value is smaller than the old one.

The disadvantage of such approach is that it will require keeping three
connections to the K8s from each Patroni Pod (previously it was two).