Get rid of kubernetes module
The kubernetes python official client contains a lot of auto-generated
code and therefore very heavy, but we need only a little fraction of it.
The naive implementation takes about 200 loc, and about half of it is
responsible for the handling of configuration files.

Disadvantage: If somebody was using the `patronictl` outside of the podi
(on his machine), it might not work anymore (depending on the environment).