Don't create permanent physical slot with name of the primary
It is a regular issue that primary is recycling WALs when one of the
replicas is down for a long time. So far there were only two solutions
for such a problem and both of them are not perfect:
1. Increase `wal_keep_segments`, but it is hard to guess the good value
2. Use continuous archiving and PITR, but it is not always possible

This PR is introducing the way to solve the problem for static clusters,
with fixed number of nodes and names that never change. You just need to
list names of all nodes in the `slots` so the primary will not remove
the slot when the node is down (not registered in DCS). Off course, the
primary will not create the permanent slot which is matching its own
name.

Close #280