Improve behavior during the etcd rolling upgrade
In dynamic environments it is common that during the rolling upgrade
etcd nodes are changing their IP addresses. If the etcd node where
Patroni is currently connected to is upgraded last, it could happen
that the cached topology doesn't contain any live node anymore and
therefore request can't be retried and totally fails, usually resulting
in demoting of the primary.

In order to partially overcome the problem, Patroni is already doing a
periodic (every 5 minutes) rediscovery of the etcd cluster topology, but
in case of very fast node rotation there was still a possibility to hit
the issue.

This commit targets to fix it. If the list of nodes exhausted, Patroni
will try to perform initial discovery via an external mechanism, like
resolving A or SRV dns records and if the new list is different from the
original, Patroni will use it as the new etcd cluster topology.

Besides that, this commit makes `_machines_cache` property pretty much
static, it will be updated only when the topology has changed and helps
to avoid concurrency issues.