Take advantage of pg_stat_wal_recevier
So far Patroni was parsing recovery.conf or querying pg_settings in
order to get the current values of recovery parameters.
On PostgreSQL earlier than 12 it could easily happen that the value of
primary_conninfo in the recovery.conf has nothing to do with reality.
Luckily for us, on PostgreSQL 9.6+ there is a pg_stat_wal_receiver view,
which contains current values of primary_conninfo and primary_slot_name.
The password field is masked through, but this is fine, because
authentication happens only during opening the connection. All other
parameters we compare as usual.

Another advantage of pg_stat_wal_recevier - it contains the current
timeline, therefore on 9.6+ we don't need to use the replication
connection trick if walreceiver process is alive.

If there is no walreceiver process available or it is not streaming
we will stick to old methods.