Sync HA loops of all pods in one cluster
There is no expire mechanism available on K8s, therefore we implement
soft leader lock, i.e. every pod is "watching" for changes of the leader
object and when there is no changes during the TTL it starts leader race.

Before we switched to LIST+WATCH approach in #1189 and #1276, we only
watched for the leader object and every time it was updated, the main
thread of the HA loop was waking up. As a result, all replica pods were
synchronized and starting the leader race more or less at the same time.

The new approach made all pods "unsynchronized" and the biggest downside
of it - it takes `ttl + loop_wait` in the worst case to detect the leader
failure.

This commit makes all pods in one cluster to sync HA loops again based on
updates of the leader object.