Detect a new timeline in the standby cluster
The standby cluster doesn't know about leader elections in the main
cluster and therefore the usual mechanisms of detecting divergences
don't work. For example it could happen that the standby cluster is
ahead of the new primary of the main cluster and must be rewound.
There is a way to know that the new timeline has been created by
checking presence of a history file in pg_wal. If the new file is
there, we will start usual procedures of making sure that we can
continue streaming or will run the pg_rewind.