Shorten and beautify history log output
when Patroni is trying to figure out the necessity of pg_rewind it could
write the content history file from the primary into the log.

The history file is growing with every failover/switchover and
eventually starts taking too many lines in the log, most of them are not
so much useful.
Instead of showing the raw data we will show only 3 lines before the
current replica timeline and 2 lines after.