Properly handle the exception raised from refresh_session
The `touch_member()` could be called from the finally block of the
`_run_cycle()`. In case if it raised an exception the whole Patroni
process was crashing.
In order to avoid future crashes we wrap `_run_cycle()` into the
try..except block and ask user to report a BUG.

Close https://github.com/zalando/patroni/issues/1529