Switch API socket to TLS after connection
The API endpoint is a threaded HTTP server, so it creates a new thread
to handle each request.  However, because it wraps its listening
socket in SSL, the ssl module insists on completing the TLS handshake
before it returns from accept(), so the entire TLS handshake happens
in the main thread.  This means that if the TLS handshake stalls for
any reason, the API endpoint becomes unresponsive.

This commit moves the wrap_socket() call so that it happens only after
each connection has been accepted, meaning that it happens in the
per-connection thread and can't affect the main one.