Defer TLS handshake until thread has started
The `SSLSocket` is immediately doing the handshake on accept.
Effectively it blocks the whole API thread if the client side doesn't
send any data. In order to solve the issue we defer the handshake until
a thread serving request has started.

The solution is a bit hacky, but thread-safe.

Close https://github.com/zalando/patroni/issues/1545