Ignore hba_file and ident_file when they match with defaults
It is possible to specify custom hba_file and ident_file in the
postgresql configuration parameters and Patroni is considering that
these files are managed externally. It could happen that locations of
these files matching with default locations of pg_hba,conf and
pg_ident.conf. In this case we will ignore custom values and fallback to
the default workflow, i.e. Patroni will owerwrite them.

Close: https://github.com/zalando/patroni/issues/1544