Bugfix: GET /cluster was showing stale member info in zookeeper
Zookpeeper implementation heavily relies on cached version of cluster
view in order to minimize the number of requests. Having stale members
information is fine for Patroni workflow because it basically relies
only on member names and tags.

The `GET /cluster` is a different case. Being exposed outside it might
be used for monitoring purposes and therefore we should show the
up-to-date members information.