Don't let Patroni mistake logical replication slots for physical replicas
In testing a PostgreSQL plugin that uses logical slots alongside Patroni,
we found that the logical slots were mistakenly showing in the Patroni REST
API output as though they were physical Patroni-managed replicas. This
wasn't causing an obvious problem in Patroni's functioning, however the
output was still incorrect. Although we did not observe such, we
believe that this may also impact failovers if Patroni were to try to
failover to the (erroneous) logical replication slot.

We tracked this behavior to the SQL statement seen here. This adds logic
to select only physical slots when considering Patroni replicas.