Fix a few possible race conditions
1. Between get_cluster() and update_leader() calls the K8s leader object
might be updated from outside and therefore the resource version will
not match (error code=409). Since we are watching for all changes, the
ObjectCache likely will have the most up-to-date version and we will
take advantage of that. There is still a chance to hit a race-condition,
but it would be smaller than before. Other DCS are free of this issue.
Etcd - update is based on the value comparison, zookeeper and consul
are relying on session mechanism. The leader race is still relying on
the version of leader object as it was during get_cluster() call.

2. If the update still failed - recheck the resource version of the
leader object and that the current node is still the leader there and
repeat the call.