Make sure cached last_leader_operation is up-to-date on replicas
Patroni is caching the cluster view in the DCS object because not all
operations require the most up-to-date values. The cached version is
valid for TTL seconds. o far it worked quite well, the only known
problem was that the `last_leader_operation` for some DCS
implementations was not very up-to-date:

* Etcd: since the `/optime/leader` key is updated right after the
'/leader` key, usually all replicas get the value from the previous HA
loop. Therefore the value is somewhere between `loop_wait` and
`loop_wait*2` old. We improve it by using the 10ms artificial sleep after
receiving watch notification from `compareAndSwap` operation on the
leader key. It usually gives enough time for the primary to update
the `/optime/leader`. On average that makes the cached version
`loop_wait/2` old.

* ZooKeeper: Patroni itself is not so much interested in most up-to-date
values of member and leader/optime ZNodes. In case of the leader race it
just reads everything from ZooKeeper, but during normal operation it is
relying on cache. In order to see the recent value on replicas they are
doing watch on the `leader/optime` Znode and will re-read it after it
was updated by the primary. On average that makes the cached version
`loop_wait/2` old.

* Kubernetes: last_leader_operation is stored in the same object as the
leader key itself and therefore update is atomic and we always see the
latest version. That makes the cached version `loop_wait/2` old on avg.

* Consul: HA loops on the primary and replicas are not synchronized,
therefore at the moment when we read the cluster state from the Consul
KV we see the last_leader_operation value that is between 0 and
loop_wait old. On average that makes the cached version `loop_wait` old.
Unfortunately we can't make it much better without performing periodic
updates from Consul, what might have negative side effects.

Since the `optime/leader` is only updated at most once per HA loop
cycle, the value stored in the DCS is usually `loop_wait/2` old on avg.
For majority of DCS implementations we could promise that the cached
version in Patroni will match the value in DCS most of the time,
therefore there is no need to make additional requests. The only
exception is Consul, but probably we could just document it, so when
someone relying on last_leader_operation value to check the replication
lag can correspondingly adjust thresholds.

Will help to implement #1599